# Optimize runtime metrics