## Summary of changes

Optimize the process information collection in the runtime metrics, on Windows.

| Method |       Mean |    Error |   StdDev | Ratio | Allocated | Alloc Ratio |
|------- |-----------:|---------:|---------:|------:|----------:|------------:|
|    Old | 5,629.8 us | 95.00 us | 88.87 us |  1.00 |    6765 B |        1.00 |
|    New |   246.3 us |  2.49 us |  2.21 us |  0.04 |         - |        0.00 |

## Reason for change

We have seen multiple cases where it appeared as a hotspot in customers who enabled the continuous profiler (I'm not saying that the CP is the cause, just that we're blind for other customers). Also, I often see dead `ProcessThread` objects in memory dumps, indicating that it's a significant source of allocations.

## Implementation details

The old version uses the .NET `Process` object to retrieve the required information. On Windows, it relies on `NtQuerySystemInformation`, which is very slow because it actually queries the information for **all** processes on the machine. To make things worse, .NET allocates `ProcessThread` objects for each thread, even though we're only interested in the count. To make worse things worse, `ProcessThread` objects have a finalizer, increasing the probability that they end up in gen 2.

This PR relies on the [process snapshotting APIs introduced in Windows 8.1](https://learn.microsoft.com/en-us/previous-versions/windows/desktop/proc_snap/process-snapshotting-portal). They're much faster because they target a single process, and use copy-on-write. The old API is kept as a fallback in case we encounter a non-transient error (arbitrarily defined as "3 consecutive errors") or if the target application uses an older version of Windows.

I had to declare separate `PSS_PROCESS_INFORMATION_32` and `PSS_PROCESS_INFORMATION_64` structures because to my surprise the `Pack` argument of `StructLayout` is 8 even on 32 bit (I naively thought it used the pointer size). It doesn't look like there is any good solution for this today, except duplicating the code: https://github.com/dotnet/runtime/issues/60573.

## Test coverage

Added a unit test for the process metrics. They do only basic validation, as anything more complex would likely flake.