@zacharycmontoya 2024-03-12
Regarding the app trimming and assembly reference snapshot diffs, this PR adds a reference to the `System.ComponentModel.Win32Exception` type, which is in the `Microsoft.Win32.Primitives`. This new reference is acceptable because this is always included in the runtime ✅ The expected files now just need to be updated appropriately